refactor(utils): change file name format of generate uid util
re #7
